refactor: 인증과 관련된 action.ts를 auth 라이브러리로 옮김
